refactor: Use `transform` utility and namedtuples for context
